requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
optuna>=3.4
shap>=0.43
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
jupyter>=1.0
nbformat>=5.0
espn_api>=0.46
